rename: ConsultingRoomController->StompController
